pyproject.toml
[project]
name = "theology-rag"
version = "0.1.0"
description = "신학 문헌 RAG — 위르겐 몰트만으로 시작하는 교부·현대 신학자 RAG 시스템"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "chromadb>=0.5.0",
    "pymupdf>=1.24.0",
    "sentence-transformers>=3.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
]

[project.scripts]
theology-rag = "theology_rag.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/theology_rag"]
